# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Stitch API Gateway",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/catalog/uom": {
      "get": {
        "tags": [
          "Catalog"
        ],
        "summary": "List Uoms",
        "operationId": "list_uoms_api_v1_catalog_uom_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/modules__catalog__application__use_cases__list_catalog_entities__UOMResponse"
                  },
                  "title": "Response List Uoms Api V1 Catalog Uom Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Catalog"
        ],
        "summary": "Register Uom",
        "operationId": "register_uom_api_v1_catalog_uom_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterUOMRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/modules__catalog__application__dto__requests__UOMResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/catalog/attributes": {
      "get": {
        "tags": [
          "Catalog"
        ],
        "summary": "List Attributes",
        "operationId": "list_attributes_api_v1_catalog_attributes_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/modules__catalog__application__use_cases__list_catalog_entities__ServiceAttributeResponse"
                  },
                  "title": "Response List Attributes Api V1 Catalog Attributes Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Catalog"
        ],
        "summary": "Define Attribute",
        "operationId": "define_attribute_api_v1_catalog_attributes_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DefineServiceAttributeRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/modules__catalog__application__dto__requests__ServiceAttributeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/catalog/offerings": {
      "get": {
        "tags": [
          "Catalog"
        ],
        "summary": "List Offerings",
        "operationId": "list_offerings_api_v1_catalog_offerings_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/modules__catalog__application__use_cases__list_catalog_entities__ServiceOfferingResponse"
                  },
                  "title": "Response List Offerings Api V1 Catalog Offerings Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Catalog"
        ],
        "summary": "Draft Offering",
        "operationId": "draft_offering_api_v1_catalog_offerings_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DraftServiceOfferingRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/modules__catalog__application__dto__requests__ServiceOfferingResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/catalog/uom/{uom_id}": {
      "put": {
        "tags": [
          "Catalog"
        ],
        "summary": "Update Uom",
        "operationId": "update_uom_api_v1_catalog_uom__uom_id__put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "uom_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Uom Id"
            }
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUOMRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/catalog/offerings/{offering_id}/attributes": {
      "post": {
        "tags": [
          "Catalog"
        ],
        "summary": "Attach Attribute",
        "operationId": "attach_attribute_api_v1_catalog_offerings__offering_id__attributes_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "offering_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Offering Id"
            }
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AttachAttributeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/catalog/offerings/{offering_id}/activate": {
      "post": {
        "tags": [
          "Catalog"
        ],
        "summary": "Activate Offering",
        "operationId": "activate_offering_api_v1_catalog_offerings__offering_id__activate_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "offering_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Offering Id"
            }
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/catalog/offerings/{offering_id}/archive": {
      "post": {
        "tags": [
          "Catalog"
        ],
        "summary": "Archive Offering",
        "operationId": "archive_offering_api_v1_catalog_offerings__offering_id__archive_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "offering_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Offering Id"
            }
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/catalog/offerings/{offering_id}": {
      "put": {
        "tags": [
          "Catalog"
        ],
        "summary": "Update Offering",
        "operationId": "update_offering_api_v1_catalog_offerings__offering_id__put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "offering_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Offering Id"
            }
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateServiceOfferingRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AttachAttributeRequest": {
        "properties": {
          "attribute_id": {
            "type": "string",
            "format": "uuid",
            "title": "Attribute Id"
          },
          "allowed_values": {
            "items": {},
            "type": "array",
            "title": "Allowed Values"
          }
        },
        "type": "object",
        "required": [
          "attribute_id",
          "allowed_values"
        ],
        "title": "AttachAttributeRequest"
      },
      "AttributeType": {
        "type": "string",
        "enum": [
          "WASTE_TYPE",
          "CONTAINER_TYPE",
          "FREQUENCY",
          "CAPACITY",
          "NUMERIC"
        ],
        "title": "AttributeType"
      },
...
      "DefineServiceAttributeRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "title": "Name"
          },
          "attribute_type": {
            "$ref": "#/components/schemas/AttributeType"
          },
          "possible_values": {
            "items": {},
            "type": "array",
            "title": "Possible Values"
          },
          "is_required": {
            "type": "boolean",
            "title": "Is Required",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "name",
          "attribute_type"
        ],
        "title": "DefineServiceAttributeRequest"
      },
...
      "DraftServiceOfferingRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "title": "Name"
          },
          "description": {
            "type": "string",
            "maxLength": 1000,
            "title": "Description"
          },
          "category": {
            "type": "string",
            "maxLength": 100,
            "title": "Category"
          },
          "default_uom_id": {
            "type": "string",
            "format": "uuid",
            "title": "Default Uom Id"
          },
          "effective_date": {
            "type": "string",
            "format": "date",
            "title": "Effective Date"
          },
          "end_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          }
        },
        "type": "object",
        "required": [
          "name",
          "description",
          "category",
          "default_uom_id",
          "effective_date"
        ],
        "title": "DraftServiceOfferingRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RegisterUOMRequest": {
        "properties": {
          "symbol": {
            "type": "string",
            "maxLength": 50,
            "title": "Symbol"
          },
          "name": {
            "type": "string",
            "maxLength": 255,
            "title": "Name"
          },
          "base_type": {
            "$ref": "#/components/schemas/UOMBaseType"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "name",
          "base_type"
        ],
        "title": "RegisterUOMRequest"
      },
...
      "UOMBaseType": {
        "type": "string",
        "enum": [
          "VOLUME",
          "WEIGHT",
          "UNIT",
          "DISTANCE",
          "TIME"
        ],
        "title": "UOMBaseType"
      },
...
      "UpdateServiceOfferingRequest": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "category": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "effective_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective Date"
          },
          "end_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          }
        },
        "type": "object",
        "title": "UpdateServiceOfferingRequest"
      },
...
      "UpdateUOMRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "UpdateUOMRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "modules__catalog__application__dto__requests__ServiceAttributeResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "ServiceAttributeResponse"
      },
      "modules__catalog__application__dto__requests__ServiceOfferingResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "ServiceOfferingResponse"
      },
      "modules__catalog__application__dto__requests__UOMResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "UOMResponse"
      },
      "modules__catalog__application__use_cases__list_catalog_entities__ServiceAttributeResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "attribute_type": {
            "type": "string",
            "title": "Attribute Type"
          },
          "is_required": {
            "type": "boolean",
            "title": "Is Required"
          },
          "possible_values": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Possible Values"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "attribute_type",
          "is_required",
          "possible_values"
        ],
        "title": "ServiceAttributeResponse"
      },
      "modules__catalog__application__use_cases__list_catalog_entities__ServiceOfferingResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "default_uom_id": {
            "type": "string",
            "format": "uuid",
            "title": "Default Uom Id"
          },
          "attributes": {
            "items": {
              "$ref": "#/components/schemas/ServiceOfferingAttributeResponse"
            },
            "type": "array",
            "title": "Attributes"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "category",
          "status",
          "default_uom_id",
          "attributes"
        ],
        "title": "ServiceOfferingResponse"
      },
      "modules__catalog__application__use_cases__list_catalog_entities__UOMResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "base_type": {
            "type": "string",
            "title": "Base Type"
          }
        },
        "type": "object",
        "required": [
          "id",
          "symbol",
          "name",
          "base_type"
        ],
        "title": "UOMResponse"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "auth/login"
          }
        }
      }
    }
  }
}